Put protocol classes in a separate file
